requirements.txt
# ── Core LLM & RAG ─────────────────────────
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
rank-bm25>=0.2.2
google-genai>=1.0.0

# ── Document Parsing ──────────────────────
pymupdf>=1.24.0
docx2txt>=0.8

# ── UI ────────────────────────────────────
streamlit>=1.38.0
streamlit-mic-recorder

# ── PDF Export ────────────────────────────
reportlab>=4.0

# ── Utilities ─────────────────────────────
python-dotenv>=1.0.0
